Split the four monolithic test files to match their packages
After the source split, four packages were each covered by a single test
file larger than some of the packages themselves — test_curriculum.py was
1,728 lines against a 1,451-line package. The repo already had the right
convention: scripts/sweep/ is a 13-module package with 13 matching
test_sweep_*.py files. These four now follow it.

  test_reporting.py      1089 -> 6 files (largest 452)
  test_curriculum.py     1728 -> 6 files (largest 462)
  test_result_bundle.py  1229 -> 4 files (largest 454)
  test_plant_contract.py  603 -> 5 files (largest 271)

Tests moved verbatim; no assertion was changed. Collected test IDs were
compared before and after for each file and are identical, except that
test_parses_serialized_booleans_strictly moved from TestStrictRecordedGate
to a new TestParseOptionalBool — it tests parse_optional_bool, not the
gate, so it now sits with the formatting module that owns it. Suite total
is unchanged at 1712 passed / 102 skipped.

Shared fixtures were extracted rather than duplicated:
- reporting_helpers.py and result_bundle_helpers.py, following the
  existing *_helpers.py convention in this directory
- conftest.py for the two actual pytest fixtures (stable_provenance,
  raptor_layers), since a fixture imported into a module reads as an
  unused import and ruff --fix would delete it

result_bundle's helpers keep their leading underscores: plant_identity,
stage_result and stage_config already occur as bare keyword-argument names
in those tests, so renaming would invite shadowing for no gain.

Three test groupings are deliberately cross-cutting and say so in their
docstrings: the SB3-availability cases (the shim's contract spans several
submodules), the plant-contract layer tests (which layers move is the
contract), and the result-bundle audit tests (whole-run integration).
